backend/requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
